requirements.txt
Flask-RESTful==0.3.8
Flask==1.1.2
Jinja2==2.11.2
Werkzeug==1.0.1
itsdangerous==1.1.0
pytz==2020.1
requests==2.23.0
python-dotenv==0.13.0